make everything thread safe, add proper testing

+++ LogComponent.Tests/LogManagerTests.cs
@@ -3,6 +3,7 @@ using LogTest.Core;
 using LogTest.Exceptions;
 using System;
 using System.IO;
+using System.Linq;
 using System.Threading;
 using Xunit;
 using Xunit.Abstractions;
@@ -11,7 +12,7 @@ namespace LogComponent.Tests
 {
     public class LogManagerTests : IDisposable
     {
-        private readonly string _tempDirectory = Path.GetTempPath() + "LogManagerTests\\";
+        private readonly string _tempDirectory = Path.Combine(Path.GetTempPath() + "LogManagerTests") + Path.DirectorySeparatorChar;
         private readonly ITestOutputHelper output;
 
         public LogManagerTests(ITestOutputHelper output)
@@ -21,10 +22,13 @@ namespace LogComponent.Tests
 
         public void Dispose()
         {
-            Directory.Delete(_tempDirectory, true);
+            if (Directory.Exists(_tempDirectory))
+            {
+                Directory.Delete(_tempDirectory, true);
+            }
         }
 
-        public bool FileContainsString(string path, string content)
+        private bool FileContainsString(string path, string content)
         {
             using (StreamReader sr = new StreamReader(path))
             {
@@ -41,12 +45,20 @@ namespace LogComponent.Tests
         public void MainLoop_ShouldCreateNewLogFileWhenDayChanges()
         {
             var mockClock = new MockClock(new DateTime(2022, 1, 1, 23, 59, 59));
+            
             var logManager = new LogManager(_tempDirectory, mockClock);
+            
+            logManager.WriteLog("Test");
+            
+            Thread.Sleep(500);
+            
             string initialLogPath = logManager.CurrentLogPath;
-
-
+            
             mockClock.Now = new DateTime(2022, 1, 2, 0, 0, 0);
-            Thread.Sleep(200);
+            
+            logManager.WriteLog("Test2");
+
+            Thread.Sleep(500);
 
             Assert.NotEqual(initialLogPath, logManager.CurrentLogPath);
         }
@@ -56,23 +68,44 @@ namespace LogComponent.Tests
         {
             var mockClock = new MockClock(new DateTime(2022, 1, 1, 23, 59, 50));
             var logManager = new LogManager(_tempDirectory, mockClock);
-            string logPath = logManager.CurrentLogPath;
 
             var log1 = "TestFirstLog";
             var log2 = "TestSecondLog";
             var log3 = "TestThirdLog";
-
-
+            
             logManager.WriteLog(log1);
             logManager.WriteLog(log2);
             logManager.WriteLog(log3);
+            
+            Thread.Sleep(500);
+            
+            string logPath = logManager.CurrentLogPath;
 
             logManager.Stop();
-            Thread.Sleep(400);
-
+            
             Assert.True(FileContainsString(logPath, log1));
             Assert.True(FileContainsString(logPath, log2));
             Assert.True(FileContainsString(logPath, log3));
+        }
+        
+        [Fact]
+        public void Stop_ShouldStopWithFlush()
+        {
+            var logManager = new LogManager(_tempDirectory);
+
+            for (int i = 0; i < 5; i++)
+            {
+                logManager.WriteLog($"Test{i}");
+            }
+            
+            logManager.StopWithFlush();
+            
+            Thread.Sleep(1000);
+            
+            for (int i = 0; i < 5; i++)
+            {
+                Assert.True(FileContainsString(logManager.CurrentLogPath, $"Test{i}"));
+            }
         }
 
         [Fact]
@@ -84,26 +117,15 @@ namespace LogComponent.Tests
             {
                 logManager.WriteLog($"Test{i}");
             }
-            logManager.Stop(false); // stopping without flush
-            Thread.Sleep(600);
 
-            Assert.NotEmpty(logManager.LogQueue);
+            logManager.Stop();
 
-        }
+            Thread.Sleep(1000);
 
-        [Fact]
-        public void Stop_ShouldStopWithFlush()
-        {
-            var logManager = new LogManager(_tempDirectory);
-
-            for (int i = 0; i < 5; i++)
-            {
-                logManager.WriteLog($"Test{i}");
-            }
-            logManager.Stop(true);
-            Thread.Sleep(600);
-
-            Assert.Empty(logManager.LogQueue);
+            var lines = File.ReadAllLines(logManager.CurrentLogPath).ToList();
+            
+            Assert.NotEqual(6, lines.Count);
+            Assert.Equal(2, lines.Count);
         }
 
         [Fact]
@@ -111,43 +133,12 @@ namespace LogComponent.Tests
         {
             var logManager = new LogManager(_tempDirectory);
 
+            output.WriteLine(_tempDirectory);
+
             logManager.Stop();
 
             Assert.Throws<LogNotAcceptedException>(() => logManager.WriteLog("Something"));
-            Assert.False(logManager.AcceptingNewLogs);
         }
-
-        [Fact]
-        public void HandleException_ShouldSetCrashLogToLogPathByDefault()
-        {
-            
-            var logManager = new LogManager(_tempDirectory);
-            logManager.WriteLog("TestWrite");
-
-            File.Delete(logManager.CurrentLogPath);
-
-            logManager.Stop();
-
-            Thread.Sleep(300);
-
-            Assert.Equal(logManager.CrashLogPath, _tempDirectory);
-        }
-
-        [Fact]
-        public void HandleException_NotOverridingSetCrashLogPath()
-        {
-            var logManager = new LogManager(_tempDirectory);
-            logManager.CrashLogPath = _tempDirectory;
-
-            logManager.WriteLog("TestWrite");
-
-            File.Delete(logManager.CurrentLogPath);
-
-            logManager.Stop();
-
-            Thread.Sleep(300);
-
-            Assert.Equal(logManager.CrashLogPath, _tempDirectory);
-        }
+        
     }
 }
